fix: found bugs during testing

+++ modules/email/src/jobs/cleanupStoredEmails.ts
@@ -1,7 +1,7 @@
 import { ConduitGrpcSdk } from '@conduitplatform/grpc-sdk';
-import { ConfigController } from '@conduitplatform/module-tools';
 import { Config } from '../config/index.js';
 import { SentEmail } from '../models/index.js';
+import { SandboxedJob } from 'bullmq';
 
 interface File {
   _id: string;
@@ -10,7 +10,7 @@ interface File {
 }
 let grpcSdk: ConduitGrpcSdk | undefined = undefined;
 
-export default async () => {
+export default async (job: SandboxedJob<{ config: Config }>) => {
   if (!grpcSdk) {
     if (!process.env.CONDUIT_SERVER) throw new Error('No serverUrl provided!');
     grpcSdk = new ConduitGrpcSdk(process.env.CONDUIT_SERVER, 'email', false);
@@ -19,7 +19,7 @@ export default async () => {
     await grpcSdk.waitForExistence('database');
     SentEmail.getInstance(grpcSdk.database!);
   }
-  const config = ConfigController.getInstance().config as Config;
+  const config = job.data.config;
   const storeInStorage = config.storeEmails.storage.enabled;
   const limit = config.storeEmails.cleanupSettings.limit;
 
@@ -45,6 +45,7 @@ export default async () => {
     const emailIdsToDelete = await SentEmail.getInstance()
       .findMany({}, '_id', undefined, limit, 'createdAt')
       .then(r => r.map(sentEmail => sentEmail._id));
+    if (emailIdsToDelete.length === 0) return;
     await SentEmail.getInstance().deleteMany({ _id: { $in: emailIdsToDelete } });
   }
 };
